requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.3.2
bcrypt==5.0.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.115.6
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jiter==0.12.0
asyncpg==0.30.0
openai==1.54.5
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.15.0
uvicorn==0.34.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
